requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
dashscope>=1.20.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
networkx>=3.2
python-louvain>=0.16
pandas>=2.1
pyarrow>=15.0
pyyaml>=6.0
python-dotenv>=1.0
streamlit>=1.35
pyvis>=0.3.2
matplotlib>=3.8
seaborn>=0.13
pytest>=8.0
pytest-mock>=3.12
jupyter>=1.0
tqdm>=4.66
